move more stuff to util
